Admit Defeat
I officially can't get the freakin' searcher test to work.

This is going to lose me a job someday, I know, but it's just not
working. I can't figure out why, and it's driving me bonkers.

I do know, for a FACT, that this works on the actual page. Since it
works on the page, I'm declaring it "Good enough" and moving onto the
actual point of this branch.